I001D: standardize decryption process and scripts
This patch separates out the decryption services and scripts
to be more generic for qcom devices.

The device-specific recovery services remain in the
init.recovery.${ro.hardware}.rc file, while the
standard decryption services are placed in
init.recovery.qcom_decrypt.rc.

This makes the decryption services portable between multiple
qcom devices.

The vendor-based decryption service binaries are moved to
vendor/bin (where they reside in stock) and the prepdecrypt
script is used to relink them to /sbin for recovery use.

This makes the script more portable for when blobs need to be
updated, without requiring manual relinking or hex-editing.

Change-Id: Ib632fd7d3c40e5ece1640e77467cba65957ced84